Jahni [Add] 푸드파워 F 추가, 에네미 피격 무시 옵션 추가

+++ Project FG/Assets/0_ProjectFG/Scripts/Data/EnemyA Data.cs
@@ -11,9 +11,9 @@ namespace JH
 
         [field: Header("에네미 A")]
         [field: Header("공격")]
-        [field: SerializeField] public int AttackDamage { get; private set; }
-        [field: SerializeField] public int AttackSpeed { get; private set; }
-        [field: SerializeField] public int AttackCoolDown { get; private set; }
+        [field: SerializeField] public float AttackDamage { get; private set; }
+        [field: SerializeField] public float AttackSpeed { get; private set; }
+        [field: SerializeField] public float AttackCoolDown { get; private set; }
         [field: Header("근접 공격 범위")]
         [field: SerializeField] public float AttackOffset { get; private set; }
         [field: SerializeField] public float AttackRadius { get; private set; }
